Initial Construct
There are a few ways to get up and running on a vue project.
There is the cdn way which is obtaining the cdn and using the
the link src modality in index.html.

There is the @vue/cli method where after install you serve the
content as you would in normal npm run serve fashion.

There is the vue ui method which I employed because I think it is
so damn cool :)

Yeah.. initial construct.